requirements.txt
fastapi==0.116.1
google-genai==1.27.0
httpx==0.28.1
pydantic==2.11.7
python-multipart==0.0.20
requests==2.32.4
starlette==0.47.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
duckdb==1.3.2
lxml==6.0.0
matplotlib==3.10.3
numpy==2.3.2
pandas==2.3.1
pillow==11.3.0
python-dateutil==2.9.0.post0
scikit-learn==1.7.1
seaborn==0.13.2
networkx==3.5
